Fix markdown glitches in generated classifiers documentation (#204)
Glitches were caused by table-of-contents logic which in absence of
title was trying to take x first characters of description, resulting in
eg. yaml snippets cut in half.

This commits avoid glitches by **not** using dot on first sentence, so
that go-swagger knows that the first sentence is a title.