Change license to GPLv3 (#1047)
### Description of change
There is a big perception issue within the community regarding the
virality of AGPLv3. The authors of this project initially chose that
license to ensure that all modifications are released back to the
community. However, we are seeing that several organizations have
blanket banned the use of AGPLv3 because of the misunderstanding of what
modifications could mean.

Until GNU comes up with a "Lesser" AGPLv3 that hopefully fulfills the
aspirations of Aperture authors, we have decided to downgrade the
license to GPLv3 for now.

##### Checklist

- [X] Documentation is changed or added

<!-- Reviewable:start -->
- - -
This change is [<img src="https://reviewable.io/review_button.svg"
height="34" align="absmiddle"
alt="Reviewable"/>](https://reviewable.io/reviews/fluxninja/aperture/1047)
<!-- Reviewable:end -->