Update readiness test in aperture test suite (#1538)
### Description of change
In this change, we check for components under `/system/readiness` to all
be ready. If any of the components are not ready, we fail the test. This
check needs to pass 2 times repeatedly before we can pass the test
because some of the OTEL components might not be ready yet.

##### Checklist

- [ ] Tested in playground or other setup
- [ ] Screenshot (Grafana) from playground added to PR for 15+ minute
run
- [ ] Documentation is changed or added
- [ ] Tests and/or benchmarks are included
- [ ] Breaking changes