Removed not-in check for StatusCode enum (#2529)
### Description of change

Removed the not-in check as by default enum sets the first field as
default and it fails the validation while uploading policy using API.

---
- Refactor: Simplified the validation rule for the
`denied_response_status_code` field in the `RateLimiter`, `Scheduler`,
`Sampler`, and `LoadRamp` messages within the `flowcontrol.proto` file.
This change enhances the maintainability of the code by reducing
complexity in the validation process.