Conditional RBAC permissions in operator and Agent Helm Chart (#2545)
Added a new flag under `operator.rbac.featureBasedRoles` which allows
users to restrict the permissions based on their installation mode of
Agent but with some restrictions.

- New Feature: Introduced a new environment variable
`APERTURE_AGENT_SIDECAR_MODE_ENABLED` to control the behavior of the
agent reconciler. When set to "true", the agent reconciler will not own
the `DaemonSet` and `Service` resources, providing more flexibility in
resource management.
- Refactor: Updated the `SetupWithManager` function in the
`AgentReconciler` struct to conditionally add ownership based on the
value of `APERTURE_AGENT_SIDECAR_MODE_ENABLED`.
- Refactor: Modified the controller setup in `main.go` to conditionally
enable the `NamespaceReconciler` controller based on the value of
`APERTURE_AGENT_SIDECAR_MODE_ENABLED`.