Reverted all go upgrades (#2570)
The go upgrade is not allowing us to update aperture on cloud due to
some weird issues of Protobuf namespace conflict which is not there for
1.20.6

- Chore: Updated Docker base images from `golang:1.20.7-bullseye` to
`golang:1.20.6-bullseye` across multiple Dockerfiles for consistency and
to align with the updated Go version.
- Chore: Adjusted cache keys in CircleCI configuration files, including
`.circleci/config.yml`, `.circleci/continue-workflows.yml`, and
`.circleci/post-release.yaml`. This change is to ensure compatibility
with the new Go version.
- Chore: Downgraded `golang` and `golangci-lint` versions in
`.tool-versions` file to `1.20.6` and `1.53.3` respectively for
compatibility reasons.