Always use etcd in GetPolicies (#2639)
GetPolicies (used by `aperturectl policies`) now always queries etcd and
cross-checks with the local state and reports policy as VALID only if
the policy is both running and exactly matches the version present in
etcd (config/api/policies). GetPolicy was altered to be consistent with
GetPolicies.

Note: The k8s-managed policies are treated differently – for them
it's expected they're not present in etcd. For consistency with
etcd-managed policies it would be nice if GetPolicies would cross-check
with state of k8s (as it cross-checks with etcd), but this commit
doesn't change this asymmetry.

UpsertPolicy was also rewritten to not depend on local state (almost,
there's a check to avoid accidentally messing with k8s-managed
policies).

Also: Load scheduler constructor now doesn't mutate the policy proto
object (so that policy in the factory now matches policy in etcd).

Resolves: #2587