Use cloud-controller component when updating controller (#2694)
We've switched from aperture-controller to cloud-controller,
and made it into a separate deployment unit. This brings us
in line with the deployment setup.

### Description of change

##### Checklist

<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
### Summary by CodeRabbit

- Refactor: Updated the description of the `controller-endpoint`
property in the `.circleci/continue-workflows.yml` configuration file
for better clarity and understanding.
- New Feature: Enhanced the `cloud-controller` application configuration
in `.opsninja.yaml`. This includes specifying image details and Docker
context for the `aperture-controller` image, and adding two Argo
manifests with their respective image prefixes. These changes improve
the cloud controller functionality.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->