Fixes in policy deserialization and hashing (#2787)
This fixes the "rpc error: code = Internal desc = grpc: error while marshaling:
string field contains invalid UTF-8" in `aperturectl policies`, which was
caused by attempt to unmarshal protowire policy as JSON, which caused the
garbage byte to end up in error message, which we put in Reason field of the
response.

Now the "unmarshal proto or JSON" logic is extracted to a function and used in
all places that interact with api/policies in etcd.

Also, this commit makes policy hashes (more) deterministic (proto marshalling
is not deterministic by default and in cases maps were used in policy (eg.
label matcher), the policy hash could change on controller restart).